// src/app/model/airport.model.ts
export interface Airport {
  id: number;
  name: string;
  city: string;
  country: string;
  iata: string;
  icao: string;
  latitude: number;
  longitude: number;
  altitude: number;
  timezone: number;
  dst: string;
  tz: string;
  type: string;
  source: string;
  elevation: number;
}

export interface CachedResponse {
  url: string;
  data: any;
}